pyproject.toml
[tool.poetry]
name = "aoc2024"
version = "1.0"
description = "AOC 2024"
authors = ["Sean Moss <smoss117@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
mypy = "^1.13.0"
scipy = "^1.14.1"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
